# Make docker CLI tests faster by mocking the call to docker auth