Before, on master
```
ok  	github.com/GoogleContainerTools/skaffold/pkg/skaffold/build/docker	46.868s
=== Slow Tests ===
37.05s	TestDockerCLIBuild
8.05s	TestDockerCLIBuild/docker_build
8.04s	TestDockerBuildError
8.03s	TestDockerBuildError/docker_file_present
7.67s	TestDockerCLIBuild/extra_env
7.55s	TestDockerCLIBuild/buildkit
6.91s	TestDockerCLIBuild/buildkit_and_extra_env
6.87s	TestDockerCLIBuild/env_var_collisions
4.44s	TestGetAllAuthConfigs

```
After this change, 
```
ok  	github.com/GoogleContainerTools/skaffold/pkg/skaffold/build/docker	1.453s
```

/cc @marlon-gamez